Skill up changes.
Added skill up over recipe level, as the prophethy foretold.
"Cleaned" the code

+++ src/map/map.cpp
@@ -995,7 +995,7 @@ int32 map_config_default()
     map_config.mount_speed_mod = 0;
     map_config.mob_speed_mod = 0;
     map_config.skillup_chance_multiplier = 1.0f;
-    map_config.craft_chance_multiplier = 1.0f;
+    map_config.craft_chance_multiplier = 1;
     map_config.skillup_amount_multiplier = 1;
     map_config.craft_amount_multiplier = 1;
     map_config.garden_day_matters = false;
@@ -1284,7 +1284,7 @@ int32 map_config_read(const int8* cfgName)
         }
         else if (strcmp(w1, "craft_chance_multiplier") == 0)
         {
-            map_config.craft_chance_multiplier = (float)atof(w2);
+            map_config.craft_chance_multiplier = atoi(w2);
         }
         else if (strcmp(w1, "skillup_amount_multiplier") == 0)
         {
